gh-archive-bigquery/requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
idna==3.11
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
